requirements.txt
# 核心依赖
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0

# 深度学习和计算机视觉
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0  # YOLO
mediapipe>=0.8.0    # 姿态检测

# Web框架
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-redis>=5.2.0     # Django Redis缓存后端

# 数据库
psycopg2-binary>=2.9.0  # PostgreSQL
redis>=4.0.0            # Redis缓存

# 数据处理
pandas>=1.3.0
scipy>=1.7.0

# 配置和序列化
PyYAML>=6.0
python-decouple>=3.6

# 异步处理
celery>=5.2.0
kombu>=5.2.0

# 监控和日志
psutil>=5.8.0  # 系统监控
loguru>=0.6.0  # 日志
watchdog>=2.1.0  # 文件监控

# 通知服务
# smtplib is built-in to Python
twilio>=7.0.0    # 短信 (可选)

# 开发和测试
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# 部署
gunicorn>=20.1.0
docker>=6.0.0

# 工具库
python-dotenv>=0.19.0
click>=8.0.0
tqdm>=4.64.0